requirement.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
construct==2.10.67
crispy==0.6.3
cycler==0.10.0
decorator==4.4.2
discord==1.0.1
discord.py==1.6.0
Django==3.1.6
django-crispy-forms==1.11.0
dnspython==2.1.0
fabio==0.11.0
google==3.0.0
h5py==3.1.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.2
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
ply==3.11
prompt-toolkit==3.0.18
protobuf==3.17.3
pyasn1==0.4.8
pycodestyle==2.6.0
pycryptodomex==3.10.1
pygame==2.0.1
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
selenium==3.141.0
silx==0.14.0
six==1.15.0
sklearn==0.0
soupsieve==2.2
sqlparse==0.4.1
tensor==0.3.6
threadpoolctl==2.1.0
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.6.3
zope.interface==5.4.0
